make it more modular
