+++ src/__init__.py
@@ -1,0 +1,14 @@
+# __init__.py
+from .database import init_database
+from .scheduler import start_scheduler
+from .data_retrieval import daily_data_retrieval
+from .cli import setup_cli, display_config_help, display_current_config
+
+__all__ = [
+    'init_database',
+    'start_scheduler',
+    'daily_data_retrieval',
+    'setup_cli',
+    'display_config_help',
+    'display_current_config'
+]
