+++ src/cli/__init__.py
@@ -1,0 +1,4 @@
+# cli/__init__.py
+from .cli import setup_cli, display_config_help, display_current_config
+
+__all__ = ['setup_cli', 'display_config_help', 'display_current_config']
