+++ src/data_retrieval/__init__.py
@@ -1,0 +1,4 @@
+# data_retrieval/__init__.py
+from .data_retrieval import daily_data_retrieval
+
+__all__ = ['daily_data_retrieval']
