+++ src/scheduler/__init__.py
@@ -1,0 +1,4 @@
+# scheduler/__init__.py
+from .scheduler import start_scheduler
+
+__all__ = ['start_scheduler']
